# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shijie Learning Platform API",
    "description": "课堂 → 练习 → 巩固 智能学习闭环平台后端。Contract-first; /openapi.json 是唯一机器契约。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/realtime/tickets": {
      "post": {
        "tags": [
          "realtime"
        ],
        "summary": "Realtime Ticket",
        "description": "Issue a short-lived one-time ticket for the WS handshake (browser cannot set\nAuthorization header on WebSocket). lesson_id must belong to the user.",
        "operationId": "realtime_ticket_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TicketOut": {
        "properties": {
          "ticket": {
            "type": "string",
            "title": "Ticket"
          },
          "ws_url": {
            "type": "string",
            "title": "Ws Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 300
          }
        },
        "type": "object",
        "required": [
          "ticket",
          "ws_url"
        ],
        "title": "TicketOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}